optimization: don't seek to 0 the first time
mitigates #105 but does not solve it.

After we open a file, we are at the beginning of the file.
So if we get a seek request for 0 and we haven't seeked yet,
we ignore it. This works around glitches caused by ffmpeg's
seeking algorithm.

However, it does not solve the case where the user seeks to
a previous item on the playlist and then a file which has already
been played comes up again.